requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.2
django-autoslug==1.9.9
django-ckeditor-5==0.2.15
django-cors-headers==4.6.0
django-jazzmin==3.0.1
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.8
google-auth==2.36.0
idna==3.10
inflection==0.5.1
isodate==0.7.2
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.0
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==24.1
pillow==11.0.0
platformdirs==4.3.6
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rpds-py==0.20.1
rsa==4.9
setuptools==75.3.0
simplejson==3.19.3
six==1.16.0
slack_sdk==3.33.3
sqlparse==0.5.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
zeep==4.3.1
